+++ extension/Changelog.md
@@ -5,7 +5,7 @@
 - Add an option to convert High-Security Padlock to Best Friend Padlock
 
 ### Bugfixes:
-
+- `ControllerIsActive()` instead of `ControllerActive` for R92
 
 ## 0.5.0
 ### Features: